add led_error_read to servo_proto

+++ src/defines.h
@@ -42,7 +42,7 @@ enum Error {
 
 struct __attribute__((__packed__)) Header {
 	int32_t cmd;
-    int8_t resp_flag;
+    int32_t resp_flag;
 	uint32_t size; 
 };
